requirements.txt
fastapi[standard]
bcrypt
psycopg2-binary
pyjwt
python-dotenv
sqlmodel
